Firmware iteration and mm on UI

+++ VC_firmware/VC_firmware/src/interface.cpp
@@ -16,7 +16,7 @@ String get_input() {
 /**
 * Check Serial for communication.
 */
-int listener(int &user_height, int &user_period) {
+int listener(float &user_height, float &user_period) {
   if (Serial.available()){
     String input = get_input();
     return validator(user_height, user_period, input);
@@ -39,7 +39,7 @@ void send_message(String channel, String mssg, float data) {
 *  }
 *
 */
-int validator(int &user_height, int &user_period, String &input) {
+int validator(float &user_height, float &user_period, String &input) {
   JsonDocument json_req;
   DeserializationError err = deserializeJson(json_req, input);
   if (err) {
